refactor(core): split agent_task_service god-file into module dir
